requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
openai==1.3.3
pydantic==2.4.2
python-multipart==0.0.6
